fix: handling virtual key pt.2
