+++ packages/pn-personagiuridica-webapp/src/components/IntegrazioneApi/ApiKeysDataSwitch.tsx
@@ -1,8 +1,6 @@
-import { MouseEvent, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { Block, Delete, MoreVert, RemoveRedEye, Sync } from '@mui/icons-material';
-import { Box, IconButton, Menu, MenuItem, Typography } from '@mui/material';
+import { Box, Chip, Typography } from '@mui/material';
 import { Row, StatusTooltip } from '@pagopa-pn/pn-commons';
 import { CopyToClipboardButton } from '@pagopa/mui-italia';
 
@@ -14,6 +12,16 @@ import {
 } from '../../generated-client/pg-apikeys';
 import { ApiKeyColumnData, ModalApiKeyView } from '../../models/ApiKeys';
 import { getApiKeyStatusInfos } from '../../utility/apikeys.utility';
+import ApiKeyContextMenu from './ApiKeyItemMenu';
+import VirtualKeyItemMenu from './VirtualKeyItemMenu';
+
+type Props = {
+  data: Row<ApiKeyColumnData>;
+  keys: BffPublicKeysResponse | BffVirtualKeysResponse;
+  type: keyof ApiKeyColumnData;
+  handleModalClick: (view: ModalApiKeyView, apiKeyId: string) => void;
+  menuType: 'publicKeys' | 'virtualKeys';
+};
 
 const isApiKeyDisactivated = (data: Row<ApiKeyColumnData>): boolean =>
   data.status !== PublicKeyStatus.Active && data.status !== VirtualKeyStatus.Enabled;
@@ -21,116 +29,7 @@ const isApiKeyDisactivated = (data: Row<ApiKeyColumnData>): boolean =>
 const setRowColorByStatus = (data: Row<ApiKeyColumnData>): string | undefined =>
   isApiKeyDisactivated(data) ? 'text.disabled' : undefined;
 
-const ApiKeyContextMenu = ({
-  data,
-  keys,
-  handleModalClick,
-}: {
-  data: Row<ApiKeyColumnData>;
-  keys: BffPublicKeysResponse | BffVirtualKeysResponse;
-  handleModalClick: (view: ModalApiKeyView, apiKeyId: string) => void;
-}) => {
-  const apiKeyId = data.id;
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-  const { t } = useTranslation(['integrazioneApi', 'common']);
-  const open = Boolean(anchorEl);
-
-  const handleClick = (event: MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  const checkIfStatusIsAlreadyPresent = (status: PublicKeyStatus): boolean =>
-    keys.items.some((key) => key.status === status);
-
-  return (
-    <Box data-testid="contextMenu">
-      <IconButton
-        onClick={handleClick}
-        size="small"
-        color="primary"
-        data-testid="contextMenuButton"
-        aria-label={t('context-menu.title')}
-        aria-controls={open ? 'context-menu' : undefined}
-        aria-haspopup="true"
-        aria-expanded={open ? 'true' : undefined}
-      >
-        <MoreVert />
-      </IconButton>
-
-      <Menu
-        data-testid="menuContext"
-        anchorEl={anchorEl}
-        open={open}
-        onClose={handleClose}
-        onClick={handleClose}
-        PaperProps={{
-          elevation: 0,
-          sx: {
-            overflow: 'visible',
-            filter: 'drop-shadow(0px 2px 8px rgba(0,0,0,0.32))',
-            mt: 1.5,
-          },
-        }}
-        transformOrigin={{ horizontal: 'right', vertical: 'top' }}
-        anchorOrigin={{ horizontal: 'right', vertical: 'bottom' }}
-      >
-        {data.status === PublicKeyStatus.Active &&
-          !checkIfStatusIsAlreadyPresent(PublicKeyStatus.Rotated) && (
-            <MenuItem
-              id="button-rotate"
-              data-testid="buttonRotate"
-              onClick={() => handleModalClick(ModalApiKeyView.ROTATE, apiKeyId)}
-            >
-              <Sync sx={{ mr: 1 }} />
-              {t('context-menu.rotate')}
-            </MenuItem>
-          )}
-        {data.status === PublicKeyStatus.Active &&
-          !checkIfStatusIsAlreadyPresent(PublicKeyStatus.Blocked) && (
-            <MenuItem
-              id="button-block"
-              data-testid="buttonBlock"
-              onClick={() => handleModalClick(ModalApiKeyView.BLOCK, apiKeyId)}
-            >
-              <Block sx={{ mr: 1 }} />
-              {t('context-menu.block')}
-            </MenuItem>
-          )}
-
-        <MenuItem
-          id="button-view"
-          data-testid="buttonView"
-          onClick={() => handleModalClick(ModalApiKeyView.VIEW, apiKeyId)}
-        >
-          <RemoveRedEye sx={{ mr: 1 }} />
-          {t('context-menu.view')}
-        </MenuItem>
-
-        {data.status !== PublicKeyStatus.Active && (
-          <MenuItem
-            id="button-delete"
-            data-testid="buttonDelete"
-            onClick={() => handleModalClick(ModalApiKeyView.DELETE, apiKeyId)}
-            sx={{ color: 'error.dark' }}
-          >
-            <Delete sx={{ mr: 1 }} />
-            {t('button.elimina', { ns: 'common' })}
-          </MenuItem>
-        )}
-      </Menu>
-    </Box>
-  );
-};
-
-const ApiKeysDataSwitch: React.FC<{
-  data: Row<ApiKeyColumnData>;
-  keys: BffPublicKeysResponse | BffVirtualKeysResponse;
-  type: keyof ApiKeyColumnData;
-  handleModalClick: (view: ModalApiKeyView, apiKeyId: string) => void;
-}> = ({ data, keys, type, handleModalClick }) => {
+const ApiKeysDataSwitch: React.FC<Props> = ({ data, keys, type, handleModalClick, menuType }) => {
   const { t } = useTranslation(['integrazioneApi']);
 
   if (type === 'name') {
@@ -167,16 +66,37 @@ const ApiKeysDataSwitch: React.FC<{
       return <></>;
     }
     const { label, tooltip, color } = getApiKeyStatusInfos(data.status, data.statusHistory);
-    return (
+    return tooltip ? (
       <StatusTooltip
         label={t(label)}
         tooltip={data.statusHistory ? tooltip : undefined}
         color={color}
       />
+    ) : (
+      <Chip
+        id={`status-chip-${t(label)}`}
+        label={t(label)}
+        color={color}
+        sx={{ cursor: 'default' }}
+        data-testid={`statusChip-${t(label)}`}
+      />
     );
   }
   if (type === 'menu') {
-    return <ApiKeyContextMenu data={data} keys={keys} handleModalClick={handleModalClick} />;
+    return menuType === 'publicKeys' ? (
+      <ApiKeyContextMenu
+        data={data}
+        keys={keys}
+        handleModalClick={handleModalClick}
+        menuType={menuType}
+      />
+    ) : (
+      <VirtualKeyItemMenu
+        data={data}
+        keys={keys as BffVirtualKeysResponse}
+        handleModalClick={handleModalClick}
+      />
+    );
   }
   return <></>;
 };
