+++ packages/pn-personagiuridica-webapp/src/components/IntegrazioneApi/PublicKeysTable.tsx
@@ -132,6 +132,7 @@ const PublicKeysTable: React.FC<Props> = ({ publicKeys, handleModalClick }) => {
                   keys={publicKeys}
                   type={column.id}
                   handleModalClick={handleModalClick}
+                  menuType="publicKeys"
                 />
               </SmartBodyCell>
             ))}
